LayoutTests/media/modern-media-controls/placard-support/placard-support-airplay-fullscreen-no-controls-expected.txt

Testing the PlacardSupport behavior when playing media through AirPlay in fullscreen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.placard is mediaController.controls.airplayPlacard
PASS mediaController.controls.children.includes(mediaController.controls.airplayPlacard) is true

PASS successfullyParsed is true

TEST COMPLETE
